requirements.txt
# Core dependencies
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0 # Satisfies torch-audio>=2.0.0
faster-whisper
whisper-timestamped==1.15.8
moviepy==1.0.3
pytubefix
ollama==0.2.0
# Removed opencv-python, replaced with opencv-contrib-python
opencv-contrib-python>=4.8.0 # For advanced face analysis (DeepFace)
numpy==1.26.4
spacy
mediapipe==0.10.13
scikit-learn>=1.3.0 # Updated from 1.2.2
librosa>=0.10.0 # Updated from no version
soundfile==0.12.1
webcolors==1.13
Pillow

# Audio/Video processing utilities
ffmpeg-python==0.2.0
pydub==0.25.1

ffmpeg-normalize==1.24.0
# General utilities
tqdm==4.67.1
requests==2.26.0
python-dotenv==0.19.0

# Development tools (optional)
ipython==8.0.0
ruff==0.5.0
pyyaml==6.0.1
typer

# New dependencies for pipeline evolution (added or updated)
sentence-transformers>=2.2.0
transformers==4.51.3
accelerate
transformers_stream_generator # Added for Qwen-VL
qwen-vl-utils[decord]
safetensors # Added for Qwen-VL model loading

pyannote.audio>=3.1.1

deepface # Added for VideoAnalysisAgent
coqui-tts # Replaced TTS for Python 3.12+ compatibility
tf-keras # Added for transformers Keras 3 compatibility
psutil # Added for memory monitoring
